[ProbeRequestFilter] Parse the radiotap header
We have to include our own definition of the radiotap header because
unfortunately there doesn't appear to be a package with the definition.

Rather than messing around with unpackaged libraries it's easier to just
include the header. We're only really interested in the length anyway,
so we can parse the rest of the captured data.